Update Lesson4Problem1.cs
The program is working, and will oass the exercise portion of this class. But I know it could be better.

+++ cit_110/week3/code/Lesson4Problem1.cs
@@ -1,4 +1,4 @@
-using System; 
+using System;
 
 namespace purchaseProgram
 {
@@ -16,15 +16,15 @@ namespace purchaseProgram
                 The program outputs the above as concatinated 
             */
             Console.WriteLine("May I have your name?");
-            string name = Console.ReadLine(); 
+            string name = Console.ReadLine();
             Console.WriteLine("What are you looking for?");
             string item = Console.ReadLine();
             Console.WriteLine("How many do you wish to purchase?");
-            int quantity = int.Parse(Console.ReadLine());
+            double quantity = double.Parse(Console.ReadLine());
             Console.WriteLine("What's a fair price?");
-            int price = int.Parse(Console.ReadLine()); 
-            int total = quantity * price; 
-            Console.WriteLine($"{name} has ordered {quantity} of {item} and it will cost ${total}"); 
+            double price = double.Parse(Console.ReadLine());
+            double total = quantity * price;
+            Console.WriteLine($"{name} has ordered {quantity} of {item} and it will cost ${total}");
         }
     }
 }
